pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = "Simple streamlit chatbot"
authors = ["Brandt Keller <brandt.keller@defenseunicorns.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain-openai = "^0.1.3"
langgraph = "^0.0.36"
langchain = "^0.1.16"
python-dotenv = "^1.0.1"
streamlit = "^1.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
